Replace sigignore with plain signal(sig ,SIG_IGN)
The sigignore function is considered deprecated on some platforms,
giving warnings during the build:

  procflow.c:109:3: warning: ‘sigignore’ is deprecated: Use the signal
  function instead [-Wdeprecated-declarations]

  parser_gram.y:1578:2: warning: ‘sigignore’ is deprecated: Use the signal
  function instead [-Wdeprecated-declarations]

Replace the use of sigignore with the equivalent signal(XX, SIG_IGN);
the signal function is already used in the code without any configure
check.

Remove the configure check for sigignore and the fallback to sigaction.